move play functionality to the timestep component
